프로젝트/server/backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
pymysql
boto3
python-dotenv
cryptography
flasgger
flask-marshmallow
marshmallow-sqlalchemy
Flask-Migrate
flask-jwt-extended
werkzeug
marshmallow
requests